Optimize the code with proper numeric values

+++ Entity Framework/9.Exam Preparation/TravelAgency/TravelAgency/Data/DataConstraints.cs
@@ -5,20 +5,21 @@ namespace TravelAgency.Data
     public static class DataConstraints
     {
         //Customer
-        public const int CustomerFullNameMinValue = 4;
-        public const int CustomerFullNameMaxValue = 60;
-        public const int CustomerEmailMinValue = 6;
-        public const int CustomerEmailMaxValue = 50;
+        public const byte CustomerFullNameMinValue = 4;
+        public const byte CustomerFullNameMaxValue = 60;
+        public const byte CustomerEmailMinValue = 6;
+        public const byte CustomerEmailMaxValue = 50;
 
         //Guide
-        public const int GuideFullNameMinValue = 4;
-        public const int GuideFullNameMaxValue = 60;
+        public const byte GuideFullNameMinValue = 4;
+        public const byte GuideFullNameMaxValue = 60;
         public const int GuideLanguageMinValue = (int)Language.English;
         public const int GuideLanguageMaxValue = (int)Language.Russian;
 
         //TourPackage
-        public const int TourPackageNameMinValue = 2;
-        public const int TourPackageNameMaxValue = 40;
-        public const int TourPackageDescriptionMaxValue = 200;
+        public const byte TourPackageNameMinValue = 2;
+        public const byte TourPackageNameMaxValue = 40;
+        public const byte TourPackageDescriptionMaxValue = 200;
+
     }
 }
